+++ CHANGELOG.md
@@ -4,9 +4,14 @@ All notable changes will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-##[1.0.0]
-### Added
+##[1.0.1]
+### Changed
+- layout on form is evened out
 
+## Fixed
+- all colors show up in wordcloud run from source (not just red)
+
+##[1.0.0]
 ### Changed
 - only includes nonzero values in the quartiles, but they remain in the wordcloud